Working decomposition package
