+++ examples/plot_ica_blind_source_separation.py
@@ -13,7 +13,7 @@ print __doc__
 
 import numpy as np
 import pylab as pl
-from scikits.learn.fastica import FastICA
+from scikits.learn.decomposition import FastICA
 
 ###############################################################################
 # Generate sample data
