+++ scikits/learn/decomposition/__init__.py
@@ -2,6 +2,6 @@
 Matrix decomposition algorithms
 """
 
-from .nmf import NMF
+from .nmf import NMF, ProjectedGradientNMF
 from .pca import PCA, RandomizedPCA, ProbabilisticPCA
 from .fastica import FastICA
