+++ scikits/learn/setup.py
@@ -17,6 +17,8 @@ def configuration(parent_package='', top_path=None):
     config.add_subpackage('cluster/tests')
     config.add_subpackage('covariance')
     config.add_subpackage('covariance/tests')
+    config.add_subpackage('decomposition')
+    config.add_subpackage('decomposition/tests')
     config.add_subpackage('feature_selection')
     config.add_subpackage('feature_selection/tests')
     config.add_subpackage('preprocessing')
